Putting tracked value calls in client.data
