+++ answer_rocket/data.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 
 from dataclasses import dataclass, field
-from typing import Optional, List, Dict
+from typing import Any, Optional, List, Dict
 from uuid import UUID
 
 import pandas as pd
@@ -19,7 +19,7 @@ from answer_rocket.graphql.schema import UUID as GQL_UUID, GenerateVisualization
     Metric, Dataset, DatasetDataInterval, Database, DatabaseSearchInput, PagingInput, PagedDatabases, \
     DatabaseTableSearchInput, PagedDatabaseTables, CreateDatasetFromTableResponse, DatasetSearchInput, PagedDatasets, \
     DatabaseKShotSearchInput, PagedDatabaseKShots, DatabaseKShot, CreateDatabaseKShotResponse, \
-    DatasetKShotSearchInput, PagedDatasetKShots, DatasetKShot, CreateDatasetKShotResponse
+    DatasetKShotSearchInput, PagedDatasetKShots, DatasetKShot, CreateDatasetKShotResponse, TrackedDimensionAttribute, TrackedDimensionValuesPage
 from answer_rocket.graphql.sdk_operations import Operations
 from answer_rocket.types import MaxResult, RESULT_EXCEPTION_CODE
 
@@ -2447,3 +2447,90 @@ class Data:
         result = self._gql_client.submit(op, mutation_args)
 
         return result.update_dataset_kshot_visualization
+    
+    def get_tracked_dimension_values(
+        self, dataset_id: UUID
+    ) -> List[TrackedDimensionAttribute]:
+        """
+        Get currently-active starred values for the calling user on a dataset,
+        grouped by dimension.
+
+        Parameters
+        ----------
+        dataset_id : UUID
+            Dataset the starred values belong to.
+
+        Returns
+        -------
+        List[TrackedDimensionAttribute]
+            One entry per dimension the caller has stars in. Each entry exposes
+            ``dimension_attribute_id``, ``dimension_name``, and ``values``
+            (currently-active starred values for that dimension).
+
+        Examples
+        --------
+        >>> attrs = max.tracked_values.get_tracked_dimension_values(
+        ...     dataset_id=uuid.UUID("...")
+        ... )
+        >>> for attr in attrs:
+        ...     print(attr.dimension_name, attr.values)
+        """
+        query_args = {"datasetId": str(dataset_id)}
+        op = Operations.query.get_tracked_dimension_values
+        result = self._gql_client.submit(op, query_args)
+        return result.get_tracked_dimension_values
+
+    def get_all_tracked_dimension_values(
+        self,
+        offset: int = 0,
+        limit: int = 100,
+        dataset_id: Optional[UUID] = None,
+        filters: Optional[dict] = None,
+        sort: Optional[list] = None,
+    ) -> TrackedDimensionValuesPage:
+        """
+        Get a flat, paginated view of tracked values. Admins see every user's
+        rows; non-admins see only their own.
+
+        Parameters
+        ----------
+        offset : int, default 0
+            Zero-based offset for paging.
+        limit : int, default 100
+            Max rows to return.
+        dataset_id : UUID, optional
+            If provided, scope results to a single dataset.
+        filters : dict, optional
+            AG-Grid filterModel JSON (per-column filters). Keys are column ids
+            (e.g. ``"userName"``, ``"value"``, ``"addedUtc"``); values follow
+            AG-Grid's text/set/date filter shapes.
+        sort : list, optional
+            AG-Grid sortModel JSON (list of ``{"colId": ..., "sort": "asc" |
+            "desc", "sortIndex": ...}`` entries).
+
+        Returns
+        -------
+        TrackedDimensionValuesPage
+            Object with ``total_count`` (count after filters but before paging)
+            and ``rows`` (the requested page of joined user/dataset rows).
+
+        Examples
+        --------
+        >>> page = max.tracked_values.get_all_tracked_dimension_values(
+        ...     dataset_id=uuid.UUID("..."),
+        ...     limit=500,
+        ... )
+        >>> print(page.total_count)
+        >>> for row in page.rows:
+        ...     print(row.user_name, row.dimension_name, row.value)
+        """
+        query_args: dict[str, Any] = {
+            "offset": offset,
+            "limit": limit,
+            "datasetId": str(dataset_id) if dataset_id is not None else None,
+            "filters": filters,
+            "sort": sort,
+        }
+        op = Operations.query.get_all_tracked_dimension_values
+        result = self._gql_client.submit(op, query_args)
+        return result.get_all_tracked_dimension_values
